Divided tests into long_run and general to overcome travis time out
