+++ hicexplorer/test/general/test_hicAggregateContacts.py
