+++ hicexplorer/test/general/test_hicCorrectMatrix.py
