+++ hicexplorer/test/general/test_hicCorrelate.py
