+++ hicexplorer/test/general/test_hicMergeMatrixBins.py
