+++ hicexplorer/test/general/test_hicPlotDistVsCounts.py
