+++ hicexplorer/test/general/test_hicPlotMatrix.py
