+++ hicexplorer/test/general/test_hicmatrix.py
